refactor: :truck: Rename mock tasks to have `mock` word at the beginning
